DbicAppCombo2: user_editable UI customizations
There are now several default visual differences applied when
'user_editable' is enabled:

  * Uses the search icon for the trigger
  * sets no_click_trigger (trigger limited to button)
  * no emtpy text
  * No match doesn't display "no matches" empty list

This mode of operation assumes the user is going to enter arbitary
values, so the search/selection functionality is made less intrusive

Note: there is also a new ra-x-form-arrow-trigger-light css class which
is a lightened version of the standard trigger. I created this just to
look slightly different, intending to use it for 'user_editable' but
decided to go with the search trigger (already build in) instead. But,
since it is already done, I left it in for possible future or end-user
use...