Default renderer for 'bigtext' now wraps long lines
Added new render func Ext.ux.RapidApp.nl2brWrap which is now used as the
default renderer for the TableSpec 'bigtext' column profile (which is
applied automatically to 'text' columns) instead of
Ext.util.Format.nl2br. This does the same thing as the standard nl2br,
but also wraps in a span with the 'ra-wrap-on' css class which sets the
CSS style 'white-space: normal'. This mainly applies to and overrides
the default ExtJS grid styles which truncates with ellipsis (...).

I think this is a better sane/default since multi-line values already
expand into mutliple lines. The expectation is already that if the user
doesn't want tall rows they should hide text columns, and wrapping long
lines is more consistent with thsi existing assumption

(set dev version 0.99501_04)