Added /rapidapp/module/printview/[path] (DirectCmp)
Added 3rd controller controller dispatch path 'printview' to the
DirectCmp controller that renders to the built-in '->printview' Module
method (alternative to '->viewport'). This support is only so-so at the
moment (there is not full Module support) but it is now available
anyway. Support for this functionality needs to be expanded in the
future...

(set dev ver 0.99501_07)