AutoPanel: extra explicit loading-indicator removal
In cases where the inner dom is not cleared when setting the body (such
as when setting an error) there is now an extra step to still try to
find and remove the loading-indicator element. The clearing step which
is ran during the course of normal rendering cannot be ran when setting
the error, and in these cases, the loading-indicator was being left
which could sometimes cause issues (same issues that applied in #46). As
a workaround, an extra step of finding and removing just the indicator
is now performed.

(set dev version 0.99501_13)