Multi-rel counts consider 'where' again (w/ warning) - fixes #95
This fixes a regression introduced in the change to using
CorrelateRelationship in 0be555eac6 which exposed an existing/known DBIC
bug. This has been addressed on a temporary basis in this commit until
the issue is resolved in DBIC.

In general, 'where' attrs in relationships shouldn't be used anyway
(extrel/coderef to build a complex ON clause is a better idea). So,
while the where attr kinda/sorta works again, a warning is now also
displayed whenever it is encountered telling the user they should to
change it to use an ON clause, with a link to GH Issue #95.

(set dev ver 0.99600_01)