RapidDbic: added more flexible 'grid_class' param
Like 'table_class' except it can be supplied at different layers of the
config to apply globally, to the model, or the individual source.

The 'table_class' param is now deprecated and replaced by 'grid_class'
within the documentation, but still works at the global level as it did
before. Any code currently using 'table_class' can substitute with
'grid_class' for identical behavior, but grid_class also now works at
additional, finer-grained levels that table_class never did.

The name 'table_class' was confusing anyway; the rationale was
originally that the class need not be a grid, but since 'grid_params'
option exists this was nullified, so now the naming is consistent and
far less prone to cause confusion

See updated POD for more info