quote_names => 1 -- NO LONGER A HARD REQUIREMENT - fixes #99
For better support for existing/older DBIC schemas/models, the
requirement that the 'quote_names' setting be enabled has been dropped,
since some code in the wild relies on hacks that do not work when
quoting is turned on.

For all new schemas generated, quote_names is still enabled by default.
The die call in AppDbicTree has been changed to a warn