Turned off setting 'storeId' in DataStore2 - fixes #100
Also in in-code comments:

 storeId was used for optimization to support re-use of stores on the
 JavaScript side via Ext.StoreMgr, but this caused caching problems
 starting with Perl 5.18. Without this setting, every componenent/tab
 will have its own, independent Ext.data.Store, and these cannot be seen
 within Ext.StoreMgr. This will lead to a larger memory footprint in the
 client, but there is no way around it at this point.

It appeared that the Ext.StoreMgr was able to smartly handle the re-use
of stores with different data for different components, as it worked
perfectly on perl 5.16 and prior. Starting with perl 5.18, the order of
data returned in the JSON changed, which made it stop working. The only
conclusion at this point I can make is that the fact that it worked
before was the actual bug.