Re-enable storeId, but append req_num + JS on destroy cleanup - #100
In the previous commit the setting of 'storeId' was disabled, however,
there are some code paths which rely on storeId to resolve/access a data
store by its id, specifically related to componenets which make use of
the 'defer_DataStore' API. The design of this case should probably be
revisited, however, in the meantime to preserve the existing API and
also fix the caching bug that is present in perl 5.18 and later, the
storeId now has the request number appended. This ensures that the same
componenent will always have a unique storeId when opened multiple
times on the client side, but getStore lookup code which is generated
within the same request will still work.

Also added extra cleanup code to the DataStorePlus javascript plugin to
lookup and remove the store by Id from Ext.StoreMgr on destroy of the
parent component. This prevents store objects from piling up in the
global StoreMgr collection in the client machine for components which
may have already been closed/destroyed.