Fix DataStore exception handler regression - #100
Changes for #100 created a regression which broke the store exception
handler. This has been fixed in this commit by moving the handler from
being set on the server side (using very very old code) to the client
side in the datastore-plus plugin (which didn't even exist at the time
the original code was written).