AppCombo: improvements to "Enter" key event behavior
AppCombo Enter key behavior is now customized and overrides some of the
Ext.combo.ComboBox logic to be more sane. The key is now ignored when a
query is in progress, and also ignored when the current value is
invalid. It also behaves accordingly with 'user_editable' mode

Also updated AppDV to call triggerBlur with its Enter event, which fixes
some long time bugs with Enter with Combo fields (see in-code comment)