Remove RapidApp::AppPageViewer (clears PPI dep) - #106
The AppPageViewer module supported viewing misc content. It was
experimental prior to the Template Controller system, which largely
replaced its functionality. However, one thing it did which has not yet
been replaced is viewing of syntax highlighted perl code. This required
a bunch of deps, including PPI, which will now be cleared from the dep
tree.

This module was not publically advertised and not in use automatically,
but was kept around for reference mainly. It can now be found in the
v0.996xx branch. Further, the way it did the viewing of perl code will
probably be handled on the client-side in the future implementation
anyway (such as with http://alexgorbatchev.com/SyntaxHighlighter/ or
http://prismjs.com)

(set dev ver 0.99699_01)