Req Catalyst::Controller::SimpleCAS 0.992 - fix Insert Image
Require new Catalyst::Controller::SimpleCAS version with added support
for forwarding to a view for JSON rendering, and this will use the
RapidApp 'RapidApp::JSON' view. This fixes Insert Image which was broken
at some point during unfactoring work. This view does special handling
to encoded the JSON response in an iframe. This is still ugly, but this
is going back to teh original behaviour and is at least way cleaner than
it was before.

Also added Devel::Confess to DevelopRequires