UNFACTOR 'ExcelTableReader' - fixes #106, fixes #110
The Excel read/parse packages were unused (except for a rouge use line
that was been removed). Only the ExcelTableWriter is used for
downloading grid data to Excel.

This cleans up the last case of multiple packages ina  single pm file
(#110) and also is where we'll declare the "phase 2" cleanup as done...
(#106)