RapidApp::Util refactor + upgr-rapidapp.pl - #74, #115
The overly complex RapidApp::Include has been unfactored and
consolidated into a simple Exporter class RapidApp::Util. In fact, all
three of these packages have been consolidated into RapidApp::Util and
disabled with a 'die' if loaded:

 * RapidApp::Include
 * RapidApp::Sugar
 * RapidApp::Functions

Additionally, the #74 conv_099_to_100.pl script has been expanded and
moved into scripts/ (i.e. will be officially installed as a script from
CPAN) and renamed:

 * upgr-rapidapp.pl

In addition to converting class names, it also now handles converting
user code that uses RapidApp::Include to the equivelent RapidApp::Util.

The idea is that this is just teh start, and that this script will be
maintained going forward to be ran against user code to upgrade it
whenever RapidApp API changes are made (which hopefully will be very
rare, but still bound to come up again in the future)