Misc internal fixes+cleanup in Tree/NavTree Modules
Implemented new 'prepare_node' method in base Tree class to support
better subclassing. This is now hooked by NavTree to apply the
nav-options per default to all nodes, including for subclasses which
implement their own fetch_nodes. This was not the original purose of
NavTree, but this is the more which is the more expected behaviour.

Also lowered the default max recurse depth from 10 to a more reasonable
and safer value of 3 (it really depends on the layout of the tree which
value is safe). If the user really wants more than this they can change
the setting