AppTree: ajax req param 'root_node' set for root node
Started custom loader class Ext.ux.RapidApp.AppTreeLoader which is now
used by Ext.ux.RapidApp.AppTree. The first thing this class does is set
the request param 'root_node' when fetching the root node. There is no
logic in RapidApp that does anything differently when this is set,
however, it is now available to other modules and is a useful data-point
in general to be available

(set dev ver 1.0010_04)