Move 'load_modules' init logic + more flexible 'navtrees' cfg
The logic to process the config option 'load_modules' has been
refactored to be passed into the rootModule in the already existing
'auto_init_modules' param. This not only simplifies and de-duplicates
this logic, but also changes the load order of these modules to occur
before the optional "main" module is loaded so that these modules are
already available for its initialization.

Also expanded the Module::Explorer (setup automatically by TabGui) to
accept {module} param in navtrees as an existing, absolute module path
instead of trying to create a new, local module.

These two changes combined allow for more flexible config, for example,
this now works:

    'RapidApp' => {
      load_modules => {
        mymod => {
          class => 'My::Module::Class',
          params => { foo => 'blah' }
        }
      }
    },

    'Plugin::RapidApp::TabGui' => {
      navtrees => [{
        module => '/mymod',
      }]
    }

This also still works:

    'Plugin::RapidApp::TabGui' => {
      navtrees => [{
        module => 'mymod',
        class => 'My::Module::Class',
        params => { foo => 'blah' }
      }]
    }

But the module is loaded as a submodule '/main/mymod' -- with the
first config, the top-level '/mymod' is used instead. This allows
for cleaner/shorter URLs

(set dev ver 1.0010_05)