Simplified and more flexible cfg for 'favicon' + misc
The favicon for the app can now be specified in the RapidApp cfg param
'default_favicon_url' as an alternative to setting the TT var within a
local_js_css.tt template file (which also still works). The benefit of
the new method is that it is also used for the '/favicon.ico' request
which will be generated by browsers when no rel="icon" link tag is set
at all, so the new setting applies not only to the ExtJS/TabGui
interface, but every other page as well.

Also added optional 'right_footer' cfg param to TabGui (still defaults
to 'title' cfg when not set)

(set dev ver 1.0010_08)