Implementing Grid Multi-Sort #128
For implementing multiple sorting ability, we need to extend the grid
for handling the changes on the header menu for the column, and also
extend the store to be able to work with the new information gathered
from the header menu.

The ExtJS Grid and the RapidApp::Module::Grid get the attribute
*use_multisort* which allows to deactivate the feature overall,
setting this to false, will prevent the ExtJS plugin for the multisort
be loaded at all.

The ExtJS Store and RapidApp::Module::StorCmp get the new attribute
*multisort_enabled*, which defines if the actual multisort is in use
for this search. The new attribute *sorters* is holding an array of
hashs which contains the *SortInfo*, exactly like ExtJS hold it.

In RapidApp::Module::StorCmp::Role::DbicLnk it was required to adapt
to the new variable *sorters* given from the ExtJS to the serverside
Store. Here this variable gets converted to the *order_by* of the DBIC
ResultSet of the search.

After every update of the *sorters* the system also adds the sorting
icons in the same way as the regular sorting mechanism of ExtJS does
it, and additionally adds a numbering to the headers in use of the
multiple sortings.

In this context it was required to extend the *Clear Sort* plugin, to
be aware of multiple sorting and makes itself invisible, if it is
enabled.

The implementation hooks into the *updateHeaderSortState* function of
the grid view to take the newly set *SortInfo* of the regular sorting
ability of ExtJS to extend *sorters*. The removing is implemented
independent but always also sets the regular ExtJS sorting to null. As
the sorting itself is happening on the server side, there was no
additional code required around the ExtJS Store to make it happen.