Implementing of "Add sort" sub menu in the header of the grid.
This shortcut will appear if you already have a sorting activated and
is a way to activate multisort and additional append the specific
field to the sorting. This is like the different direction then to
click directly on the menu for sorting.

The implementation is pretty straightforward adding a new menu which
is only shown when required via a check on *beforeMenuShow* in the
multisort plugin.

This still misses icons for this submenu items.