Switching to Chained Dispatcher is very important to be part of the general Catalyst Ecosystem, and especially where the future of Catalyst will be a Chained only world. The given structure showed up problems on several level and made the integration of regular normal Catalyst Controller into the infrastructure of RapidApp impossible. The biggest problem here is the element of Catalyst to follow always the rule of the last set Dispatcher that fits the request, which is in this case **always** RapidApp, as RapidApp gets injected into the Components of the Catalyst Application **after** the user components are loaded. This can't be turned around, else RapidApp wouldn't know anything of the user application, and can't react in an intelligent way, which is part of the base concepts of RapidApp, to be intelligent reacting to the needs.
The switch to Chained was more or less straightforward, it only
needed a bit of *unwiring* about the default handler given through
RapidApp::Template::Controller::Dispatch, which was programmed to
claim the root namespace at all cost. This behaviour is now replaced
with a check for a User given root controller, and if there is none
he installs a Path action default handler. It sadly has to be path
to not disturb any given RapidApp application. In a future version
we will also add generation of Chained default handler. In the
dream vision, the system would autodetect what is required and
installs the proper default handler (like if the user has Path
Controllers we need to install a Path default handler).

No existing behaviour of RapidApp should be hit. This commit message
is also relevant for the 2 previous commit messages, which were
steps to reach this target.