Require latest 'base' pkg 2.18 (dist.ini) - #143
This fixes a very strange problem which cropped up with base 2.15
*after* fcd34a88a50 when trying to start the RA-InfoSite app (at
https://github.com/RapidApp/RA-InfoSite/commit/ca8c2d2bd1). It would
throw this error with base 2.15:

 Can't locate object method "mk_classdata" via package "DBIx::Class::Schema" ...

base 2.18 has been the latest version for 3 years though, and its very
strange what could have triggered this. In any case, the latest version
which works is now required. Perhaps DBIC should be requiring this as
well?