AppDV: load xtemplate only on 'content'
The template loading of AppDV is now only called at 'content' endpoint
and _NOT_ on every request. This fixes bugs related to accessing sub
module paths under the AppDV module, such as accessing RESTful rel
paths. When called on every request, not only is it needlessly expensive,
but it triggers failures such as "Record Id not supplied" when it
tries to access req_Row, when an AppDV is being uses as a RowPg, as
is the case with the new DbicRowDV module

Set dev ver 1.0401_06