Multi-rel col render now uses (modified) with-inline-icon
Tweaked the 'with-inline-icon' css class to be truly/better "inline"
for the case of a 16x16 icon and 16px line-height, and switched the
mutli-rel column render/markup to use it instead of 'with-icon' which
is block. This allows for multi-rel columns to work natively/properly
in inline contexts, such as AppDV templates, but still work properly
in grids, etc.

It may be a good idea to replace most uses of 'with-icon' with
'with-inline-icon' but so far, the multi-rel case is all that has been
switched...