Start new column profile 'cas_img' + new logics (IN PROG)
In progress commit...

Started new column profile 'cas_img' and started new behaviour for
cas-image-field to store values in the same manner as 'cas_link' which
stores only a checksum/name instead of full-blown <img> tags. Added
new renderer 'Ext.ux.RapidApp.renderCasImg' which is able to display
these values as <img> tags and also to parse the old, full-blown <img>
values.

This is mostly working, but there are remaining details regarding
sizing which still need to be figured out...

IN PROGRESS