NavCore: ordering bug fixes + handling node drop correctly
Fixed issues regarding ordering "My Views" properly and also now
handle node_move (dd) when the target is a top-level container. The
node is properly ordered as the last item within the target for this
case. This not only now behaves correctly, but it should also fix an
exception which is thrown by Pg (when Pg is used for the coreschema)
when this bad case was hit.

(set dev ver 1.0401_13)