Relative path fallback/dispatch for special controllers
Module dispatch will now perform a redispatch for bad end-points when
they match one of the known, top-level controller paths (such as
'simplecas' and 'assets'). This allows for simple relative paths
in client-side markup (img tags, etc) w/o fuss and for them to be
able to work regardless of the mount_url.

Additionally, those components which generate markup links, namely
HtmlEditor which inserts images, now insert the relative URL (note
that the fallback logic is not needed for normal cases for this to
work, since the tab context is already relative, regardless of the
mount_url, however, other cases, like module/direct urls, do need
the new fallback dispatch for those inserted images to be able to
render

(set dev ver 1.0401_14)