ADD NEW MAGIC PATH DISPATCH KEYWORD '_ra-rel-mnt_' - #153
If the now reserved keyword/string '/_ra-rel-mnt_/' appears anyplace
in the url path in any request, we munge it (within Middleware) and
strip everything in the path *up to that point* before sending into
Catalyst/RapidApp dispatch. The reason this is being done is to
provide an efficient mechanism for generated markups such as <img>
'src', css urls, <link> tags, etc, to reference paths (such as
simplecas and asset urls) w/o needing to know the current mount_url
and, more importantly, be able to remain valid if the mount_url is
changed after the fact. This is essentially a means to supply an
absolute url path but the form of a *relative* url path from the
perspective of the browser. This is a better way to handle this case
than from within the module dispatch because it avoids the
associated unnecessary overhead (which can vary from module to
module) and is also even more flexible, to also work via locally-
defined controller actions which internally re-dispatch to modules.

This is a second take on this problem which was already handled
within module dispatch (see _maybe_special_path_redirect logic) but
which I was never happy with because its ugly. This previous
solution does not conflict in any way with the new _ra-rel-mnt_
option, however, it will probably be unfactored and removed, if for
no other reason, keeping the code less confusing. (however, if this
is done, already generated markup in the wild will have to be
considered, but i do not think there is enough of it out there to
justify keeping the code, and the possible failure scenario is very
minor)

Internal locations which relied on referencing relative urls for
generating markup, such as the HtmlEditor insert image, have been
changed to reference the new _ra-rel-mnt_ relative paths.

This change is still considered EXPERIMENTAL and needs more testing
before release...

(set dev ver 1.0502_03)