COLUMN SUMMARIES NOW AVAILABLE ON VIRTUAL COLUMNS - fixes #93
Grid Column Summaries are now available by default for all types of
columns, including virtual and multi-rel columns, as well as single rel
columns with virtual display columns. This has been structurally
possible for a long time, but disabled by rule since before other
significant refactoring, such as unfactor of special HAVING logic
(see #51) and other changes. Since that time, this has been much
easier to do.

Besides simply removing the rule which disallowed column summaries (
'no_summary' column property) some special value handling needed to be
added to handle the ref structure of virtual columns which has been
done, as well as cleanup and simplification of related code and now
dead codepaths.

Also did some minor changes to allow renderers for multi-rel columns to
behave better when rendered as a named summary func, and added new
column profile 'multirel' which is used internally set to multi-rel
column opts for the summary funcs (as number) as well as the other
multi-rel column opts in a cleaner manner.

The only type of column for which column summaries are still disabled
is m2m, which are a special case and probably make no sense to have a
column summary at all, however, this may still revisit in the future...