DataStorePlus: Add common Ajax response access
The last response abject and decoded data is now tracked in a common
location for all requests, read and write, which is are handled
differently by the native ExtJS code. A reference to the 'lastResponse'
is now saved in the proxy object, and a reference to the 'lastJsonData'
is now saved in the store.

See in-code comments for more details