AppDV: new 'resData' available to tpl + refresh_on_save option
Added a mechanism to access the decoded json data of the last response
from within the template. For instance, this would provide access to
the value 'results' returned from the server (to inline the value):

 {[this.resData.results]}

Also added new ext param 'refresh_on_save' (different from the generic
'reload_on_save' which is common to all datastore modules) to refresh
the template after any write operation. This allows for refresh of the
template after an update (for instance in teh case of data changes
which are separate from the individual row being changed)