New attribute trait 'ExtProp' (EXPERIMENTAL)
Added new role class for use as an attribute trait to flag as an Ext
config param. This has been attempted in the past, but this new
experiment actually handles calling apply_extconfig itself and also
automatically translates Bool values into \0 and \1 (JSON true/false)

The idea behind this trait is that declaring Moose attributes will
automatically be available to the ExtJS component, and no longer
have to call apply_extconfig() manually

The first attribute which has been setup to utilize this is the new
'refresh_on_save' option just added to AppDV. It is now a Moose
attribute even though it is only processed in JavaScript, and there
is no apply_extconfig call. Assuming no issues crop up, the plan
will be wholesale refactoring of relevent attributes to use this