ADD DATASTORE COLUMN PARAM 'documentation' w/ prelim (forms) handling
Added/started new DataStore column param 'documentation' to provide
an official place to record general-purpose human-readable info about
a given column. This is a relatively simple concept which has never
been done previously in order to avoid constraining a given app into
any specific style. This has always been possible, but previously
required custom implementations. But, after revisiting the topic (and
also because of design decisions which have happened since then) I
decided that this is a common-enough, useful need that it ought to be
supported as a built-in. (note: the reason the name 'documentation'
was chosen for the param is because there is a built-in of the same
name for Moose attrs, although column params are not currently
driven by Moose, but may in the future. But the reason for the name
is purely convention)

The first built-in handing for this value, which has been added in
this commit, is StoreCmp add and edit forms. Columns which have
'documentation' set render it via the (already existing)
Ext.ux.FieldHelp plugin, which displays the content directly under
field in grey, 9px font. This may change in the future, and the
paradigm will be expanded. Next up a method to view from Grids will
be added.

Part of the reason for this very conservative, slow implementation
is because this is a new pubic (very public) API and so its needs
to stay as broad as possible (which is in fact its not been done at
all until now, which just takes the cause of being careful too far)

(set dev version 1.0601_05)