MultiFilter + Custom Header pop-up dialog constrain to grid/tab - #6
The Ext.Window based pop-up dialogs of MultiFilter and Grid Custom
Header now are constrained to their parent grid. This is helpful so
the user can now navigate to other, unrelated tabs without having
to close the browser-wide dialog.

Also fixed a long-time ExtJS CSS bug so that the style of the windows
could stay the same as they used to (this is explained in detail in
comments). It was making this change to constrain these windows which
allowed me to realize what was going on with this CSS bug and finally
fix it