Column 'documentation' in grid hmenu + smartRenderTo + misc - #6
Renamed the grid "Custom Header" plugin to "Column Properties" which
now both allows changing the header as well as display the value of
the column 'documentation' if set.

Also battled with some more ExtJS CSS bugs related to elements nested
with girds (last time it was borders, this time its scrolling
breaking). I decided to overcome any possible future, yet unknown
bugs for rendering windows (#6, which is the thing that made this
come up) by adding a new conmstructor property (via override) to
Ext.Window called 'smartRenderTo' which does the same thing as the
existing 'renderTo' property but detects if target is a grid and
jumps up one level (which is still well inside the tab and inside the
boundary which would would care about, but escapes the x-grid CSS
space which is so problematic.

Went ahead and converted the MultiFilter, Batch Modify and Column
Properties dialogs to use this new API, and also added add and edit
windows for DataStores (for when they are configured as window and
not tab) to be constrained using this as well.