AppDV: add a few px to the field size from data size
When the dimensions of the field are calculated based on the
dimensions of the inline value, 6 extra pixels are given to the
height (and 10 to the width for good measure). This helps display
scenarios such as when the stylesheet sets the font-size to match,
since text fields lose padding at the top and bottom. This change
allows the field to be able to display the same font-size w/o crop
of the text, and is a small enough increase to not cause meaningful
layout problems...