Fix toolbar breakage with autoload_added_record - ref #21, fixes #161
Tracked down the sizing problem described in #161 to calls to
syncSize() which were added in 5ba9014a7097a202e1bb3c9ef25de12183447f25 for #21.

According to the commit this was added in order to properly handle
toolbar overflows, but I beleive this was not actually correct at
the time, since the next commit also described fixing this problem
in a much more complete way (ea4a707226cbbdb660fade0f32d31cb25c25ae6c)

The syncSize() calls probably should have been removed at that time,
as they don't appear to be needed and are responsible for the breakage
in #161. I tested the overflow functionality after removing the lines
and it appears to be working correctly.

(set dev ver 1.0601_08)