DataStorePlus: Formalized 'firstload' event
Now is really called always after the store 'load' event, but only
once. It was originally implemented for one narrow case, but now it
is a full-blown event called with an expected argument list. Note:
this is still considered part of the internal/private API, but there
is no reason end-user/developer code couldn't call it (its just not
documented anyplace outside of the code itself)