Fix click/event regression from 025a7911c36d6782e9cb152a219ee3cf27691e62
Within the first commit in which the 'ra-async-box' feature was added,
the option stopEvent:true was added to the click listener in the
Template Controller js code, which breaks nav clicks within templates.
I'm not sure why I did this, I think it was just a test which slipped
through. It shouldn't have anything to do with the ra-async-box