AppDV: added option 'refresh_on_hash_change' + 'hashval' tpl variable
The new attr refresh_on_hash_change can be set to true to cause the
client-side template to refresh whenever the window.location.hash
changes (via Ext.History hooks). Also, the value of the hash can also
now be accessed via {[this.hashval]} in the template. This is the
same as {[window.location.hash]}, which was already available, but
it strips the '#' from the front of the string.

(set dev ver 1.0601_12)