Virtual Column 'set_function' now imply ->update (if makes dirty)
update() is now automatically called after calling set_function of
Virtual Columns *if* the row has been left dirty. This allows calling
methods to set other column values (i.e. $row->col1('blah')) from
within the set_function and having it take effect w/o having to call
->update() manually. This is just the more expected behavior. The
reason this hasn't been done in the past is because of the fact that
set functions are actually called *after* the standard update() and
are expected to handle whatever changes are needed in an "external"
manner. However, automatically handling this case doesn't detract
from that and it the more expected behavior (and I can't imagine any
case where the row is left with dirty columns and the user would
actually want to discard them). None of this gets in the way of the
existing behavior (i.e. the user is still free to call ->update
themself and it will work the same as it used to; they just don't
have to now)

(set dev ver 1.0601_14)