Add sugar func 'throw_prompt_ok' (high-level CustomPrompt wrapper)
New sugar function 'throw_prompt_ok' can be called from within the
context of an active request to prompt the user with a custom message
and standard 'Ok' or 'Cancel' buttons. The request is repeated with
the user's answer, which is then returned from the function (button
text is returned, so if they clicked 'Ok' this is what is returned)

This is just a sugar wrapper around the CustomPrompt system which has
been around for a long time and supports a huge array of call patterns.
This suger just makes it dirt simple to use for this common case.

See in-code comments for additional details including e.g. usage

(set dev ver 1.0601_15)