AppDV: improved handling for allow_add/allow_edit flags
AppDV now applies CSS classes to properly reflect UI elements
accoridng to underlying editable state. This prevents cases where
a column has allow_edit set to false from rendering an edit icon
which makes the column appear editable (even though clicking it would
already not do anything). AppDV now also understands the difference
between allow_edit and allow_add, and if allow_edit is false but
allow_add is true, it will do the right thing.

However, not all cominations/scenarios are yet handled in AppDV to the
same degree as in the Grid, and work on this is still in-progress
(for instance, a column missing from creatable_colspec is not handled
like if it was set with allow_add => 0, which it should). This is
due to deficiencies lower in the stack which have been planned to be
addressed during the long-planned refactor of the DataStore/DbcLink
sub-system.